OP-451-refactored - Changing the entity files organization
