+++ src/Controller/ShowCartsAction.php
@@ -11,11 +11,11 @@ declare(strict_types=1);
 namespace BitBag\SyliusMultiCartPlugin\Controller;
 
 use BitBag\SyliusMultiCartPlugin\Entity\CustomerInterface;
-use BitBag\SyliusMultiCartPlugin\Repository\OrderRepositoryInterface;
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Customer\Context\CustomerContextInterface;
 use Symfony\Component\HttpFoundation\Response;
+use Tests\BitBag\SyliusMultiCartPlugin\Application\src\Repository\OrderRepositoryInterface;
 use Twig\Environment;
 
 final class ShowCartsAction
