--- src/Resources/config/resources.yml
@@ -1,12 +1,0 @@
-sylius_order:
-    resources:
-        order:
-            classes:
-                model: BitBag\SyliusMultiCartPlugin\Entity\Order
-                repository: BitBag\SyliusMultiCartPlugin\Repository\OrderRepository
-
-sylius_customer:
-    resources:
-        customer:
-            classes:
-                model: BitBag\SyliusMultiCartPlugin\Entity\Customer
