+++ tests/Application/src/Entity/Order.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\BitBag\SyliusMultiCartPlugin\Application\src\Entity;
+
+use BitBag\SyliusMultiCartPlugin\Entity\OrderInterface;
+use BitBag\SyliusMultiCartPlugin\Model\OrderTrait;
+use Sylius\Component\Core\Model\Order as BaseOrder;
+
+class Order extends BaseOrder implements OrderInterface
+{
+    use OrderTrait;
+}
