+++ tests/Application/src/Resources/config/doctrine/Customer.orm.xml
@@ -1,14 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="BitBag\SyliusMultiCartPlugin\Entity\Customer" table="sylius_customer">
-
-        <field name="activeCart" type="integer" nullable="true" column="active_cart"/>
-
+    <mapped-superclass name="Tests\BitBag\SyliusMultiCartPlugin\Application\src\Entity\Customer" table="sylius_customer">
+        <field name="activeCart" type="integer" nullable="true" column="active_cart" />
     </mapped-superclass>
-
 </doctrine-mapping>
