+++ tests/Application/src/Resources/config/doctrine/Order.orm.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                                    http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+    <mapped-superclass name="Tests\BitBag\SyliusMultiCartPlugin\Application\src\Entity\Order" table="sylius_order">
+        <field name="cartNumber" type="integer" nullable="true" column="cart_number" />
+    </mapped-superclass>
+</doctrine-mapping>
